# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AddProjectMemberRequest:
    properties:
      role_in_project:
        type: string
      user_id:
        type: integer
    type: object
...
  dto.CreateProjectRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: invalid request body
        type: string
      message:
        example: request failed
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  dto.ProjectData:
    properties:
      created_at:
        type: string
      description:
        example: Project dau tien
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Project Alpha
        type: string
      owner_id:
        example: 1
        type: integer
      updated_at:
        type: string
    type: object
  dto.ProjectListPayload:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.ProjectData'
        type: array
      pagination:
        $ref: '#/definitions/dto.SwaggerPaginationResponse'
    type: object
  dto.ProjectListSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.ProjectListPayload'
      message:
        example: get projects success
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.ProjectMemberData:
    properties:
      email:
        example: b@example.com
        type: string
      joined_at:
        type: string
      name:
        example: Nguyen Van B
        type: string
      role_in_project:
        example: member
        type: string
      user_id:
        example: 2
        type: integer
    type: object
  dto.ProjectMemberListPayload:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.ProjectMemberData'
        type: array
      pagination:
        $ref: '#/definitions/dto.SwaggerPaginationResponse'
    type: object
  dto.ProjectMemberListSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.ProjectMemberListPayload'
      message:
        example: get project members success
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.ProjectMemberSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.ProjectMemberData'
      message:
        example: add project member success
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.ProjectSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.ProjectData'
      message:
        example: create project success
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  dto.SimpleSuccessResponse:
    properties:
      message:
        example: delete success
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.SwaggerPaginationResponse:
    properties:
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      total:
        example: 25
        type: integer
      total_pages:
        example: 3
        type: integer
    type: object
...
  dto.UpdateProjectRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: RESTful API cho hệ thống Task Management
  title: Task Management API
  version: "1.0"
paths:
...
  /projects:
    get:
      consumes:
      - application/json
      description: Lấy danh sách project
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProjectListSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List projects
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: Tạo project mới
      parameters:
      - description: Create project request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ProjectSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create project
      tags:
      - Projects
  /projects/{id}:
    delete:
      consumes:
      - application/json
      description: Xóa project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SimpleSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete project
      tags:
      - Projects
    get:
      consumes:
      - application/json
      description: Lấy chi tiết project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProjectSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get project detail
      tags:
      - Projects
    put:
      consumes:
      - application/json
      description: Cập nhật project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update project request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProjectSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update project
      tags:
      - Projects
  /projects/{id}/members:
    get:
      consumes:
      - application/json
      description: Lấy danh sách thành viên trong project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProjectMemberListSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List project members
      tags:
      - Project Members
    post:
      consumes:
      - application/json
      description: Thêm thành viên vào project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Add project member request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddProjectMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ProjectMemberSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add project member
      tags:
      - Project Members
  /projects/{id}/members/{userId}:
    delete:
      consumes:
      - application/json
      description: Xóa thành viên khỏi project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SimpleSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove project member
      tags:
      - Project Members
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"